backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
asyncpg
alembic
pydantic
pydantic[email]
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
aiofiles
jinja2
pyppeteer
sentry-sdk
pytest
pytest-asyncio
httpx
aiosqlite
